add delete view and pricefield to integer

+++ books/models.py
@@ -1,11 +1,12 @@
 from django.db import models
 from django.urls import reverse
 
+
 class Book(models.Model):
     title = models.CharField(max_length=200)
     author = models.CharField(max_length=200)
     content = models.TextField()
-    price = models.DecimalField(max_digits=5, decimal_places=2)
+    price = models.PositiveIntegerField()
 
     def __str__(self):
         return f'{self.author}: {self.title}'
